requirements.txt
acres==0.5.0
adjustText==1.3.0
bosdyn-api==4.1.1
bosdyn-client==4.1.1
bosdyn-core==4.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
Deprecated==1.2.18
durationpy==0.10
etelemetry==0.3.1
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
grpcio==1.71.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.8
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.1.0
protobuf==6.30.1
prov==2.1.1
psutil==7.0.0
puremagic==1.30
py-cpuinfo==9.0.0
pydot==4.0.1
PyJWT==2.10.1
PyMuPDF==1.26.3
pynmea2==1.19.0
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzed==1.3.0
rdflib==7.1.4
requests==2.32.3
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.166
ultralytics-thop==2.0.14
urllib3==2.3.0
vg==2.0.0
wrapt==1.17.2
